Replace bitsandbytes with torchao for cross-platform quantization
- Replace bitsandbytes dependency with torchao in pyproject.toml
- Refactor quantize_module() to use torchao.quantization.quantize_()
  with int4_weight_only and int8_weight_only schemes
- Add new QuantType enum values (INT4_WEIGHT_ONLY, INT8_WEIGHT_ONLY)
  while keeping INT8/NF4 as deprecated aliases for backward compat
- Add --quantization CLI arg (replaces --quant_type, kept as deprecated)
- Add --compile-block flag for torch.compile(mode='max-autotune') on CUDA
- Force float16 dtype throughout for cross-platform network safety
  (bfloat16 crashes Apple Silicon nodes lacking hardware support)
- Enforce explicit .to(torch.float16) at all network serialization
  boundaries (server output, client input, forward/backward paths)
- Remove all bitsandbytes imports and bnb-specific LoRA wrappers
  from peft.py (torchao keeps nn.Linear with quantized tensor subclasses)
- Update block size estimation for new quantization types

+++ src/agentgrid/client/inference_session.py
@@ -163,7 +163,7 @@ class _ServerInferenceSession:
                 runtime_pb2.ExpertRequest(
                     uid=self.uid,
                     tensors=[
-                        serialize_torch_tensor(tensor.to(proto.dtype).cpu(), proto.compression)
+                        serialize_torch_tensor(tensor.to(torch.float16).cpu(), proto.compression)
                         for tensor, proto in zip(input_tensors, inference_schema)
                     ],
                     metadata=MSGPackSerializer.dumps(request_metadata),
